refactor: restructure the folders and files a bit
