+++ client/src/index.js
@@ -1,7 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './index.css';
-import './pages/sports-grid/SportsGrid.css';
 import 'bootstrap/dist/css/bootstrap.min.css';
 
 import App from './App';
